fixed new post
